requirements.txt
openai
z3-solver
requests
python-dotenv
pandas
matplotlib